+++ sympy/printing/numpycode.py
@@ -1,0 +1,4 @@
+from .pycode import PyCodePrinter
+
+class NumPyCodePrinter(PyCodePrinter):
+    pass
