+++ sympy/printing/scipycode.py
@@ -1,0 +1,4 @@
+from .numpycode import NumPyCodePrinter
+
+class SciPyCodePrinter(NumPyCodePrinter):
+    pass
